backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
webcolors==1.11.1
djoser==2.1.0
Pillow==9.0.0
django-filter==23.5
gunicorn==20.1.0